refactor(mcp): split Tables tools into their own file with shared schemas
Move the 5 FlowFuse Tables tools out of teams.js into tools/tables.js,
and extract their repeated input/output field schemas (teamId,
databaseId, tableName, schemaName, database shape, count/meta records)
into a new tool-schemas/tables.js module, since the same fields are
reused across both the input and output schemas.

Also drops the plan-gated-feature caveat from tool descriptions and
switches remaining .passthrough() calls to the non-deprecated .loose().
